+++ src/service/event.ts
@@ -634,15 +634,18 @@ export function createScreeningEventSeriesFromCOA(params: {
         dateMvtkBegin: params.filmFromCOA.dateMvtkBegin
     };
 
-    const acceptedPaymentMethod: factory.paymentMethodType[] = [
-        factory.paymentMethodType.Account,
-        factory.paymentMethodType.Cash,
-        factory.paymentMethodType.CreditCard,
-        factory.paymentMethodType.EMoney
-    ];
+    // const acceptedPaymentMethod: string[] = [
+    // ];
+    let unacceptedPaymentMethod: string[] | undefined;
 
     if (coaInfo.flgMvtkUse === '1') {
-        acceptedPaymentMethod.push(factory.paymentMethodType.MovieTicket);
+        // acceptedPaymentMethod.push(factory.paymentMethodType.MovieTicket);
+    } else {
+        if (!Array.isArray(unacceptedPaymentMethod)) {
+            unacceptedPaymentMethod = [];
+        }
+
+        unacceptedPaymentMethod.push(factory.paymentMethodType.MovieTicket);
     }
 
     return {
@@ -691,7 +694,8 @@ export function createScreeningEventSeriesFromCOA(params: {
             project: { typeOf: params.project.typeOf, id: params.project.id },
             typeOf: factory.offerType.Offer,
             priceCurrency: factory.priceCurrency.JPY,
-            acceptedPaymentMethod: acceptedPaymentMethod
+            // ...(Array.isArray(acceptedPaymentMethod)) ? { acceptedPaymentMethod: acceptedPaymentMethod } : undefined,
+            ...(Array.isArray(unacceptedPaymentMethod)) ? { unacceptedPaymentMethod: unacceptedPaymentMethod } : undefined
         },
         additionalProperty: [
             {